NMS-10454: Improve unit test, and additional debugging
Set the charset to UTF-8 when reading the packet capture for the 36-byte
offset test.
Compare the individual bytes of the identifier addresses, instead of the
2-byte strings to which they decode.
Use org.xbill.DNS.Address when resolving the hostname on an event. Works
around DNS issues on Windows hosts when resolving 'localhost'.
Add some debug statements around the hostname resolution process.