Fix for Bug NMS-10609
The MIB Compiler is unable to parse certain MIBs.

A few welfomed MIBs were not properly parsed, due to a problem with
the necessary loop required by the third party library we use to
parse MIB dependencies.

All the existing tests and the new test added here (NMS10609Test)
are passing without issues. Without the fix, the new test fails.