HZN-1556: Collect steps in a dedicated layer for caching and remove the indermediate image
Rearranged the build steps in a Multi-Stage build in such a way, all external repetitive and
expensive YUM install tasks are in a base stage. This stage is cached and these get invalidated
as soon we change a) RUN commands, build arguments and our OpenJDK base image. The build arguments
which are used in OCI labels are moved to the bottom, cause they are changed very frequently and
invalidate the following layers with every build.