NMS-12483: Add docker buildx for releases as multi arch and single OCI
Enhance our bild job to build for master a multi-architecture OCI with a manifest for arm/v7, arm64 and amd64 when we
make a release and push it directly to the registry.

To run same build instruction and remove Bash script complexity, I've migrated the our build scripts to a single
Makefile. When run make it creates the OCI file as local build for the single x86_64 architecture as OCI file. The
version number as docker tag is fetched from the root pom.xml. For a release we create additionally a OCI file for
x86_64 as artifact file.