NMS-8484: Added pointer-like-index property extenders for SNMP collector
PointerLikeIndexPropertyExtender is a generic extender which uses the value
of an attribute on a target resource (e.g. cpmCPUTotalPhysicalIndex) as a
foreign-key on which it then "joins" a source resource to retrieve the value
of a nominated string attribute (e.g. entPhysicalName) on that source resource.
This class has the same name as one contributed in an abandoned PR by
Jean-Marie Kubek, but in the end I wrote my own implementation.

InterfaceSnmpPropertyExtender is a specific extender for "joining" resources
of type interfaceSnmp against the value of an attribute on a target attribute
which is a value of ifIndex. The separate implementation is necessary because
interfaceSnmp resources are indexed not by their ifIndex, but by a calculated
ifLabel.